attributes long adding method in memory add_old process changing in memory attributes  and constructor add one more constructor for in Segment for free and old (see this for gui

+++ MemoryManagement/src/memorymanagementAlgorithm/Void.java
@@ -108,8 +108,8 @@ public class Void {
         return void_vector.size();
     }
 
-    public int get_total_size() {
-        int size = 0;
+    public long get_total_size() {
+        long size = 0;
         for (int i = 0; i < void_vector.size(); i++) {
             size += void_vector.get(i).getLimit();
         }
